requirements/requirements.txt
click==8.1.7
colorama==0.4.6
numpy==2.1.2
Pillow==11.0.0
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyqt5-tools==5.15.9.3.3
python-dotenv==1.0.1
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
setuptools==65.5.0
pytest==7.4.0
